requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
appnope==0.1.4
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
exceptiongroup==1.2.2
executing==2.2.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.2
gast==0.4.0
google-auth==2.38.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
idna==3.10
imageio==2.37.0
importlib_resources==6.5.2
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.35.0
jax==0.4.26
jax-metal==0.1.0
jaxlib==0.4.26
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==2.12.0
kiwisolver==1.4.8
lazy_loader==0.4
Levenshtein==0.27.1
libclang==18.1.1
lightning-utilities==0.14.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.2.0
mpmath==1.3.0
multidict==6.3.2
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.4
oauthlib==3.2.2
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.15.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
pluggy==1.5.0
prompt_toolkit==3.0.50
propcache==0.3.1
protobuf==4.25.6
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.1
pyparsing==3.2.3
pytesseract==0.3.13
pytest==8.3.5
python-dateutil==2.9.0.post0
python-Levenshtein==0.27.1
pytorch-lightning==2.5.1
pytz==2025.2
PyWavelets==1.8.0
PyYAML==6.0.2
pyzmq==26.4.0
RapidFuzz==3.13.0
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rsa==4.9
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
stack-data==0.6.3
sympy==1.13.3
tabulate==0.9.0
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.1
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.12.0
tensorflow-metal==1.0.0
termcolor==3.0.1
threadpoolctl==3.6.0
tifffile==2025.3.30
tomli==2.2.1
torch==2.2.2
torchaudio==2.2.2
torchmetrics==1.7.0
torchvision==0.17.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.14.1
yarl==1.19.0
